change all longs to int to avoid confusion

+++ section-2/step-01/src/main/java/com/carmanagement/agentic/agents/CleaningAgent.java
@@ -36,7 +36,7 @@ public interface CleaningAgent {
             String carMake,
             String carModel,
             Integer carYear,
-            Long carNumber,
+            Integer carNumber,
             String rentalFeedback,
             String cleaningFeedback);
 }
