+++ section-2/step-05/multi-agent-system/src/main/java/com/carmanagement/agentic/agents/DispositionFeedbackAgent.java
@@ -44,7 +44,7 @@ public interface DispositionFeedbackAgent {
             String carMake,
             String carModel,
             Integer carYear,
-            Long carNumber,
+            Integer carNumber,
             String carCondition,
             String rentalFeedback,
             String cleaningFeedback,
